feat: move `HookContext` into `hook` instead of `fmt`
