+++ packages/libs/error-stack/src/hook/mod.rs
@@ -1,3 +1,5 @@
+pub(crate) mod context;
+
 use alloc::vec::Vec;
 
 use crate::{
